add Node interface
